Endre UtsettelseDialog til connected container
