+++ src/app/components/utsettelseSkjema/utsettelseSkjema.less
@@ -1,7 +1,3 @@
-.utsettelseSkjemaDialog {
-	width: 100%;
-	max-width: 25em;
-}
 .utsettelseSkjema {
 	padding: 2rem 1rem 0 1rem;
 	&__hovedknapp {
