+++ src/app/styles/planlegger.less
@@ -1,5 +1,5 @@
 @import '../../shared/styles/shared.less';
-@import '../components/utsettelseDialog/utsettelseDialog.less';
+@import '../components/utsettelseSkjema/utsettelseSkjema.less';
 @import './moduloverstyring.less';
 
 p {
@@ -54,4 +54,9 @@ p {
 		right: 0;
 		top: -2px;
 	}
+}
+
+.utsettelseSkjemaDialog {
+	width: 100%;
+	max-width: 25em;
 }
